requirements.txt
azure-storage-blob==12.14.1
elasticsearch==7.17.6
elasticsearch_dsl==7.4.0
frontend==0.0.3
PyMuPDF==1.19.0
Flask==2.1.0
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
geopy==2.2.0
numpy==1.21.5
pandas==1.4.4
python_docx==0.8.10
sentence_transformers==2.2.2
spacy==3.4.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
python-dotenv
elastic-apm[flask]
blinker
gunicorn==20.1.0
Authlib==1.2.0
werkzeug==2.0.3
